add samples, update ws mock
